azure_deploy/pyproject.toml
[project]
name = "azure_deploy"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "langchain-openai (>=0.3.21,<0.4.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "azure-search-documents (>=11.5.2,<12.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "website @ file:///Users/borisdev/workspace/azure-fastapi-nextjs/backend",
]
package-mode = true
packages = [{include = "*", from = "azure_deploy"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
website = {develop = true}
